feat(charms): complete offensive + leech/crit on-hit charms (phase B5c)
Expands the on-hit charm dispatcher (player hits assigned monster):
- Overpower: physical, min(8% target HP, 5% player max HP).
- Overflux: physical, min(8% target HP, 2.5% player max mana).
- Cripple: paralyzes the monster 10s.
- Vamp / Void: life / mana leech for chance%% of the damage dealt.
- Low / Savage: crit-style bonus physical burst (approximated — the port has no
  crit pipeline to modify).

Charm chance is now float32 so Vamp/Void fractional leech (1.6%%, 0.8%%) survives.
Elemental charms (B5b) unchanged.

+++ internal/charms/charms.go
@@ -58,10 +58,10 @@ type Charm struct {
 	Description string
 	Category    uint8
 	Type        uint8
-	DamageType  int     // COMBAT_* value from the datapack (creatures_definitions.hpp)
-	Percent     float64 // percent of the reference health/mana used as damage
-	Chance      [3]uint16
-	Points      [3]uint16 // charm-point (major) or minor-echo (minor) cost per tier
+	DamageType  int       // COMBAT_* value from the datapack (creatures_definitions.hpp)
+	Percent     float64   // percent of the reference health/mana used as damage
+	Chance      [3]float32 // trigger chance (or leech %) per tier; fractional for Vamp/Void
+	Points      [3]uint16  // charm-point (major) or minor-echo (minor) cost per tier
 	Effect      uint16
 	CastSound   uint16
 	ImpactSound uint16
